docs: Update variable in Readme.org's lambda function
Using this plugin's instruction for ssh causes Emacs flycheck-mode to warn that
`dirtrackp` shown in the Readme.org doc is unavailable/undefined anywhere. Doing
further investigation I found that this variable is not defined or used in
ssh.el. I decided update the variable used in the Readme.org's lambda function
instruction from `dirtrackp` to `shell-dirtrackp` which is what is actually
defined and used in ssh.el just in-case this might help someone else in the
future.